requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
APIFlask==2.1.0
apispec==6.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==2.1.1
click==8.1.7
flasgger==0.9.7.1
Flask==3.0.1
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.1.0
frozenlist==1.4.1
github==1.2.7
gunicorn==20.0.4
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.4
marshmallow==3.20.2
mistune==3.0.2
multidict==6.0.4
packaging==23.2
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
webargs==8.4.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
